feat: made CacheService Generic
